ops/ci/buildbot/configuration/requirements.txt
pyyaml==3.12
pytest==3.0.4
buildbot==0.9.2
PyOpenSSL
